rdmdownload to new file rdmdownload.c
