+++ src/Makefile.am
@@ -22,11 +22,11 @@ bin_PROGRAMS = remotedebugger
 remotedebuggerincludedir = $(includedir)/rrd
 remotedebuggerinclude_HEADERS = rrdCommon.h rrdInterface.h
 
-remotedebugger_SOURCES = rrdMain.c rrdEventProcess.c rrdJsonParser.c rrdDeepSleep.c rrdRunCmdThread.c rrdCommandSanity.c rrdExecuteScript.c rrdMsgPackDecoder.c rrdInterface.c
+remotedebugger_SOURCES = rrdMain.c rrdEventProcess.c rrdJsonParser.c rdmdownload.c rrdRunCmdThread.c rrdCommandSanity.c rrdExecuteScript.c rrdMsgPackDecoder.c rrdInterface.c
 remotedebugger_CFLAGS = -I$(top_srcdir)/include/rrd -I$(PKG_CONFIG_SYSROOT_DIR)${includedir}/trower-base64/ -I$(PKG_CONFIG_SYSROOT_DIR)${includedir}/rbus/ $(CJSON_CFLAGS)
 AM_LDFLAGS="-lpthread -lrdkloggers -lmsgpackc -ltrower-base64 -lwebconfig_framework -lrbus -lsecure_wrapper"
 if IARMBUS_ENABLE
-remotedebugger_SOURCES += rrdIarmEvents.c 
+remotedebugger_SOURCES += rrdIarmEvents.c rrdDeepSleep.c
 remotedebugger_CFLAGS += -I$(PKG_CONFIG_SYSROOT_DIR)${includedir}/rdk/iarmbus/ -I$(PKG_CONFIG_SYSROOT_DIR)${includedir}/rdk/iarmmgrs/rdmmgr -I$(PKG_CONFIG_SYSROOT_DIR)${includedir}/rdk/iarmmgrs-hal -DIARMBUS_SUPPORT
 AM_LDFLAGS += "-lIARMBus -lrfcapi -ltr181api"
 endif
